(alldocs) Conventions.lpdoc included from BasicLang
It was included in the introduction of the reference manual, before the
tour. It talks about syntax, predicates, modes, types, etc. It reads
surprisingly better in the "Ciao basic language" part! That is the first
part that really requires it, and the text connects the basic part with
the next one (assertions).

Src-commit: e3a5c6ab070226b89bedbb78c9a268743457b9f6